part_panshin/requirements.txt
hydra-core==1.3.2
torch-ema==0.3
bitsandbytes==0.35.0
accelerate==0.20.3
albumentations==1.4.8
opencv-python==4.10.0.82
scikit-learn==1.5.0
timm==1.0.3
tqdm==4.66.4
torch==2.3.1
torchvision==0.18.1
kaggle==1.6.14
numpy==1.26.4